Updated counting bloom filter to use byte counters instead of ushort to decrease memory usage.

+++ Maybe/BloomFilter/CountingBloomFilter.cs
@@ -2,13 +2,17 @@
 
 namespace Maybe.BloomFilter
 {
+    /// <summary>
+    /// A bloom filter modified to store counters and allow elements to be removed from the collection.
+    /// </summary>
+    /// <typeparam name="T">The type of item to be stored in the collection</typeparam>
     public class CountingBloomFilter<T> : BloomFilterBase<T>
     {
-        private readonly ushort[] _collectionState;
+        private readonly byte[] _collectionState;
 
         private CountingBloomFilter(int arraySize, int numHashes) : base(arraySize, numHashes)
         {
-            _collectionState = new ushort[arraySize];
+            _collectionState = new byte[arraySize];
             for (var i = 0; i < _collectionState.Length; i++)
             {
                 _collectionState[i] = 0;
@@ -16,10 +20,10 @@ namespace Maybe.BloomFilter
         }
 
         /// <summary>
-        /// Creates a new bloom filter with appropriate bit width and hash functions for your expected size and error rate.
+        /// Creates a new counting bloom filter with appropriate bit width and hash functions for your expected size and error rate.
         /// </summary>
-        /// <typeparam name="T">The type of item to be held in the bloom filter</typeparam>
-        /// <param name="expectedItems">The maximum number of items you expect to be in the bloom filter</param>
+        /// <typeparam name="T">The type of item to be held in the counting bloom filter</typeparam>
+        /// <param name="expectedItems">The maximum number of items you expect to be in the counting bloom filter</param>
         /// <param name="acceptableErrorRate">The maximum rate of false positives you can accept. Must be a value between 0.00-1.00</param>
         /// <returns>A new bloom filter configured appropriately for number of items and error rate</returns>
         public static CountingBloomFilter<T> Create(int expectedItems, double acceptableErrorRate)
@@ -33,16 +37,16 @@ namespace Maybe.BloomFilter
         }
 
         /// <summary>
-        /// Adds an item to the bloom filter
+        /// Adds an item to the counting bloom filter
         /// </summary>
         /// <param name="item">The item which should be added</param>
         public void Add(T item) => DoHashAction(item, hash => _collectionState[hash]++);
 
         /// <summary>
-        /// Checks if this bloom filter currently contains an item
+        /// Checks if this counting bloom filter currently contains an item
         /// </summary>
         /// <param name="item">The item for which to search in the bloom filter</param>
-        /// <returns>False if the item is NOT in the bloom filter. True if the item MIGHT be in the bloom filter.</returns>
+        /// <returns>False if the item is NOT in the counting bloom filter. True if the item MIGHT be in the counting bloom filter.</returns>
         public bool Contains(T item)
         {
             var containsItem = true;
@@ -51,10 +55,10 @@ namespace Maybe.BloomFilter
         }
 
         /// <summary>
-        /// Removes an item from the bloom filter
+        /// Removes an item from the counting bloom filter
         /// </summary>
         /// <param name="item">The item to remove</param>
-        /// <returns>True if the bloom filter might contain the item and the item was removed. False otherwise.</returns>
+        /// <returns>True if the counting bloom filter might contain the item and the item was removed. False otherwise.</returns>
         public bool Remove(T item)
         {
             if (!Contains(item)) return false;
